+++ confiture-web-app/src/components/audit/CriteriumNotApplicableAccordion.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { useIsOffline } from "../../composables/useIsOffline";
+import Tiptap from "../ui/Tiptap.vue";
 import LazyAccordion from "./LazyAccordion.vue";
-import MarkdownHelpButton from "./MarkdownHelpButton.vue";
 
 defineProps<{ id: string; comment: string | null }>();
 
@@ -17,31 +17,14 @@ const title = "Commentaire";
 <template>
   <LazyAccordion :title="title" disclose-color="var(--background-default-grey)">
     <!-- COMMENT -->
-    <div class="fr-input-group fr-mb-1w">
-      <label class="fr-label fr-sr-only" :for="`criterum-comment-field-${id}`">
-        {{ title }}
-      </label>
-      <textarea
-        :id="`criterum-comment-field-${id}`"
-        :value="comment ?? ''"
-        class="fr-mt-0 fr-input"
-        rows="5"
-        :disabled="isOffline"
-        :aria-describedby="`markdown-notice-${id}`"
-        @input="
-          $emit('update:comment', ($event.target as HTMLTextAreaElement).value)
-        "
-      ></textarea>
-    </div>
-
-    <MarkdownHelpButton :id="`markdown-notice-${id}`" />
+    <p :id="`criterum-comment-field-${id}`" class="fr-label fr-sr-only">
+      {{ title }}
+    </p>
+    <Tiptap
+      :model-value="comment"
+      :labelled-by="`criterum-comment-field-${id}`"
+      :disabled="isOffline"
+      @update:model-value="$emit('update:comment', $event)"
+    />
   </LazyAccordion>
 </template>
-
-<style scoped>
-.markdown-notice {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-</style>