+++ confiture-web-app/src/components/ui/Tiptap.vue
@@ -7,6 +7,7 @@ import {
 import CodeBlockLowlight from "@tiptap/extension-code-block-lowlight";
 import DropCursor from "@tiptap/extension-dropcursor";
 import { Heading, type Level } from "@tiptap/extension-heading";
+import Image from "@tiptap/extension-image";
 import Link from "@tiptap/extension-link";
 import Typography from "@tiptap/extension-typography";
 import StarterKit from "@tiptap/starter-kit";
@@ -16,23 +17,31 @@ import js from "highlight.js/lib/languages/javascript";
 import ts from "highlight.js/lib/languages/typescript";
 import html from "highlight.js/lib/languages/xml";
 import { common, createLowlight } from "lowlight";
-import { onBeforeUnmount, ShallowRef } from "vue";
+import { Markdown } from "tiptap-markdown";
+import { onBeforeUnmount, ShallowRef, watch } from "vue";
 
+import { useUniqueId } from "../../composables/useUniqueId";
 import TiptapButton from "./TiptapButton.vue";
 
 export interface Props {
   modelValue?: string | null;
   editable?: boolean;
   labelledBy?: string | null;
+  disabled?: boolean;
+  editorSize?: "sm" | "lg";
 }
 
 const props = withDefaults(defineProps<Props>(), {
   modelValue: "",
   editable: true,
-  labelledBy: null
+  disabled: false,
+  labelledBy: null,
+  editorSize: "sm"
 });
 
 const emit = defineEmits(["update:modelValue"]);
+
+const uniqueId = useUniqueId();
 
 // Define needed heading levels
 const displayedHeadings = [4, 5, 6] as Array<Level>;
@@ -84,12 +93,16 @@ function setLink() {
 }
 
 // Editor attributes to create an accessible textarea
-const editorAttributes: any = {
-  "aria-describedby": "tiptap-description",
-  rows: "10",
-  "aria-multiline": "true",
-  role: "textbox"
-};
+const editorAttributes: any = props.editable
+  ? {
+      "aria-describedby": "tiptap-description",
+      "aria-multiline": "true",
+      role: "textbox",
+      class: `tiptap--${props.editorSize}`
+    }
+  : {
+      class: "tiptap--not-editable"
+    };
 
 if (props.labelledBy) {
   editorAttributes["aria-labelledby"] = props.labelledBy;
@@ -153,7 +166,9 @@ const extensions: Extensions = [
   Typography.configure({
     openDoubleQuote: "« ",
     closeDoubleQuote: " »"
-  })
+  }),
+  Markdown.configure({ linkify: true }),
+  Image
 ];
 
 if (props.editable) {
@@ -167,7 +182,7 @@ const editor = useEditor({
   editorProps: {
     attributes: editorAttributes
   },
-  editable: props.editable,
+  editable: props.editable && !props.disabled,
   content: getContent(),
   extensions,
   onUpdate({ editor }) {
@@ -176,19 +191,35 @@ const editor = useEditor({
   }
 }) as ShallowRef<Editor>;
 
+watch([() => props.editable, () => props.disabled], ([editable, disabled]) => {
+  editor.value.setEditable(editable && !disabled);
+});
+
 onBeforeUnmount(() => {
   editor.value?.destroy();
+});
+
+defineExpose({
+  focusEditor: () => {
+    editor.value.commands.focus();
+  }
 });
 </script>
 
 <template>
   <div
     class="tiptap-container"
-    :class="editable ? 'tiptap-container--editable' : null"
+    :class="{
+      'tiptap-container--not-editable': !editable,
+      'tiptap-container--disabled': disabled
+    }"
   >
-    <p id="tiptap-description" class="fr-sr-only">
-      Éditeur de texte riche, vous pouvez utiliser le format Markdown ou bien
-      utiliser les raccourcis clavier.
+    <p
+      v-if="editable"
+      :id="`tiptap-description-${uniqueId}`"
+      class="fr-sr-only"
+    >
+      Éditeur de texte riche
     </p>
     <ul v-if="editable" class="tiptap-buttons">
       <li>
@@ -199,7 +230,7 @@ onBeforeUnmount(() => {
               switch-off-label="Retirer le gras"
               icon="bold"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleBold()"
+              :disabled="!editor?.can().toggleBold() || disabled"
               :pressed="editor?.isActive('bold')"
               @click="editor.chain().focus().toggleBold().run()"
             />
@@ -210,7 +241,7 @@ onBeforeUnmount(() => {
               switch-off-label="Retirer l’italique"
               icon="italic"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleItalic()"
+              :disabled="!editor?.can().toggleItalic() || disabled"
               :pressed="editor?.isActive('italic')"
               @click="editor.chain().focus().toggleItalic().run()"
             />
@@ -221,7 +252,7 @@ onBeforeUnmount(() => {
               switch-off-label="Ne pas barrer le texte"
               icon="strikethrough"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleStrike()"
+              :disabled="!editor?.can().toggleStrike() || disabled"
               :pressed="editor?.isActive('strike')"
               @click="editor.chain().focus().toggleStrike().run()"
             />
@@ -233,7 +264,8 @@ onBeforeUnmount(() => {
               :icon="`h-${i + 1}`"
               :is-toggle="true"
               :disabled="
-                !editor?.can().toggleHeading({ level: hLevel as Level })
+                !editor?.can().toggleHeading({ level: hLevel as Level }) ||
+                disabled
               "
               :pressed="editor?.isActive('heading', { level: hLevel })"
               @click="
@@ -255,7 +287,7 @@ onBeforeUnmount(() => {
               switch-off-label="Éditer le lien"
               icon="link"
               :is-toggle="true"
-              :disabled="!editor?.can().setLink({ href: 'test' })"
+              :disabled="!editor?.can().setLink({ href: 'test' }) || disabled"
               :pressed="editor?.isActive('link')"
               @click="setLink"
             />
@@ -271,8 +303,9 @@ onBeforeUnmount(() => {
               icon="list-unordered"
               :is-toggle="true"
               :disabled="
-                !editor?.can().toggleBulletList() &&
-                !editor?.can().toggleOrderedList()
+                (!editor?.can().toggleBulletList() &&
+                  !editor?.can().toggleOrderedList()) ||
+                disabled
               "
               :pressed="editor?.isActive('bulletList')"
               @click="editor.chain().focus().toggleBulletList().run()"
@@ -285,8 +318,9 @@ onBeforeUnmount(() => {
               icon="list-ordered"
               :is-toggle="true"
               :disabled="
-                !editor?.can().toggleBulletList() &&
-                !editor?.can().toggleOrderedList()
+                (!editor?.can().toggleBulletList() &&
+                  !editor?.can().toggleOrderedList()) ||
+                disabled
               "
               :pressed="editor?.isActive('orderedList')"
               @click="editor.chain().focus().toggleOrderedList().run()"
@@ -302,7 +336,7 @@ onBeforeUnmount(() => {
               switch-off-label="Ne pas définir comme citation"
               icon="quote-line"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleBlockquote()"
+              :disabled="!editor?.can().toggleBlockquote() || disabled"
               :pressed="editor?.isActive('blockquote')"
               @click="editor.chain().focus().toggleBlockquote().run()"
             />
@@ -313,7 +347,7 @@ onBeforeUnmount(() => {
               switch-off-label="Ne pas définir comme passage de code"
               icon="code-view"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleCode()"
+              :disabled="!editor?.can().toggleCode() || disabled"
               :pressed="editor?.isActive('code')"
               @click="editor.chain().focus().toggleCode().run()"
             />
@@ -324,7 +358,7 @@ onBeforeUnmount(() => {
               switch-off-label="Ne pas définir comme bloc de code"
               icon="code-block"
               :is-toggle="true"
-              :disabled="!editor?.can().toggleCodeBlock()"
+              :disabled="!editor?.can().toggleCodeBlock() || disabled"
               :pressed="editor?.isActive('codeBlock')"
               @click="editor.chain().focus().toggleCodeBlock().run()"
             />
@@ -346,33 +380,70 @@ onBeforeUnmount(() => {
   position: relative;
   background-color: var(--background-alt-grey);
   border-radius: 0.5rem 0.5rem 0 0;
-  padding: 0.5rem 0.75rem;
-  border: 0 solid var(--border-plain-grey);
-  border-bottom-width: 1px;
-  min-height: 30rem;
+  border: 1px solid var(--border-plain-grey);
+  border-bottom: 0;
+  box-shadow: inset 0 -2px 0 0 var(--border-plain-grey);
+
+  /* Override bg color in dark mode to avoid same color as wrapper */
+  @media (prefers-color-scheme: dark) {
+    background-color: var(--background-contrast-grey);
+  }
 }
 
+.tiptap-container--not-editable {
+  padding: 0;
+  background-color: transparent;
+  border: none;
+  box-shadow: none;
+}
+
+.tiptap-container--disabled:hover {
+  cursor: not-allowed;
+}
+
+.tiptap-container--disabled .tiptap * {
+  color: var(--text-disabled-grey);
+}
+
+.tiptap {
+  padding: 0.5rem 0.75rem;
+
+  &.tiptap--not-editable {
+    min-height: 0;
+  }
+
+  &.tiptap--sm {
+    min-height: 10rem;
+  }
+
+  &.tiptap--lg {
+    min-height: 24rem;
+  }
+}
+
+/* TODO: include commented properties only in editable mode */
 .tiptap img {
-  cursor: pointer;
+  /* cursor: pointer; */
   display: block;
   height: auto;
   max-width: 100%;
   border-radius: 0.25rem;
   margin: 0.5rem 0 2rem;
-  padding: 0.5rem;
+  /* padding: 0.5rem; */
 }
 
 .tiptap:focus,
 .tiptap:focus-visible {
-  outline-width: 3px !important;
-  outline-offset: 0 !important;
+  outline-width: 2px !important;
+  outline-offset: -2px !important;
 }
 
-@media (hover: hover) and (pointer: fine) {
+/* TODO: include only in editable mode */
+/*@media (hover: hover) and (pointer: fine) {
   .tiptap img:hover {
     box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
   }
-}
+}*/
 
 .tiptap p {
   vertical-align: middle;
@@ -409,6 +480,41 @@ onBeforeUnmount(() => {
   width: var(--icon-size);
 }
 
+/* Update DSFR ol ::marker styles:
+  Nested <ol>
+  1.
+    1.1
+      1.1.1
+    1.2
+  2.
+
+  <ol> nested in <ul>
+  - 
+    1.
+    2.
+      1.
+      2.
+  - 
+*/
+.tiptap {
+  ol,
+  ul {
+    counter-reset: section;
+  }
+
+  ol li {
+    --ol-content: counters(section, ".") ". ";
+
+    counter-increment: section;
+  }
+
+  ul > li > ol li {
+    --ol-content: counter(section) ". ";
+
+    counter-increment: section;
+  }
+}
+
 .tiptap blockquote p {
   font-size: 1.25rem;
   font-weight: 700;
@@ -432,14 +538,11 @@ onBeforeUnmount(() => {
   scrollbar-width: none; /* Firefox */
   -ms-overflow-style: none; /* Internet Explorer 10+ */
   position: sticky;
+  background-color: inherit;
   top: 0;
   z-index: 1;
-  background-color: var(--background-alt-grey);
-  margin-block-end: 0.5rem;
-  padding-block: 0.25rem;
-  width: calc(100% + 1.5rem);
-  transform: translateX(-0.75rem);
-  padding-inline: 0.75rem;
+  padding: 0.5rem 0.75rem;
+  border-radius: 0.5rem 0.5rem 0 0;
 }
 
 .titptap-buttons::-webkit-scrollbar {
@@ -499,12 +602,13 @@ onBeforeUnmount(() => {
   background-color: var(--background-alt-grey-hover);
 }
 
-.tiptap-selection,
+/* TODO: include only in editable mode */
+/* .tiptap-selection,
 .ProseMirror-selectednode {
   outline-style: dotted;
   outline-width: 2px;
   outline-color: var(--dsfr-outline);
-}
+} */
 
 .ProseMirror-widget {
   opacity: 0.5;